refactor: move advertising client to module repository
